requirement.txt
python-rofi
py-notifier
